Implement westworld logging for Compatibility API.
Bug: 136794938
Bug: 135010838
Test: Not yet as it needs CP'ing to internal to do this.

+++ cmds/statsd/src/anomaly/AnomalyTracker.cpp
@@ -232,7 +232,7 @@ void AnomalyTracker::declareAnomaly(const int64_t& timestampNs, const MetricDime
 
     // TODO: This should also take in the const MetricDimensionKey& key?
     android::util::stats_write(android::util::ANOMALY_DETECTED, mConfigKey.GetUid(),
-                               mConfigKey.GetId(), mAlert.id());
+                               mConfigKey.GetId(), (int64_t)mAlert.id());
 }
 
 void AnomalyTracker::detectAndDeclareAnomaly(const int64_t& timestampNs,
